Refactor: Move form fields and material components to a new components directory
- Deleted app_form_fields.dart, app_material.dart, app_page.dart, dashboard_metric_grid_view.dart, and related files from resources.
- Created new components for app buttons, dropdown fields, form fields, material, page, and dashboard metric grid view.
- Updated imports in various files to reflect the new component structure.
- Ensured all functionalities remain intact after restructuring.
