+++ lib/views/home/components/quiz_category_card.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:quizzia/models/ui_models.dart';
 import 'package:quizzia/resources/app_colors.dart';
-import 'package:quizzia/resources/app_material.dart';
+import 'package:quizzia/components/app_material.dart';
 import 'package:quizzia/views/home/components/quiz_settings_dialog.dart';
 
 class QuizCategoryCard extends StatelessWidget {
@@ -26,7 +26,7 @@ class QuizCategoryCard extends StatelessWidget {
           },
         );
       },
-      child: Ink(
+      child: Padding(
         padding: const EdgeInsets.all(24),
         child: Column(
           crossAxisAlignment: CrossAxisAlignment.center,
